new structure. Added blueprints Inventory and Auth.
