+++ app/auth/forms.py
@@ -1,21 +1,7 @@
-
-from flask import flash
-from flask_wtf import FlaskForm
 from wtforms import SubmitField, PasswordField, StringField
 from wtforms.validators import Email, DataRequired
+from app.forms import BaseForm
 
-class BaseForm(FlaskForm):
-
-    __abstract__ = True
-    
-    def flash_errors(self):
-        """Flashes form errors"""
-        for field, errors in self.errors.items():
-            for error in errors:
-                flash(u"Error in the %s field - %s" % (
-                    getattr(self, field).label.text,
-                    error
-                ), 'error')
 
 # Define the login form (WTForms) for /login and /singup
 class LoginForm(BaseForm):
