+++ app/auth/templates/auth/sign_up.html
@@ -18,6 +18,6 @@ Sing up
         {{ form.password2(class="form-control", id="password2", placeholder="Confirm password" ) }}
     </div>
     <br />
-    {{ form.submit(class="btn btn-primary") }}
+    {{ form.submit(class="btn btn-primary", onclick="return validatePasswords()") }}
 </form>                 
 {% endblock %}
