+++ app/inventory/templates/inventory/filter.html
