+++ app/inventory/templates/inventory/home.html
