+++ app/inventory/templates/inventory/inventory.html
