+++ app/models.py
@@ -1,0 +1,11 @@
+from app import db
+from sqlalchemy.sql import func
+
+
+class Base(db.Model):
+
+    __abstract__ = True
+
+    id = db.Column(db.Integer, primary_key=True)
+    date_created = db.Column(db.DateTime, default=db.func.now())
+    date_modified = db.Column(db.DateTime, default=db.func.now(), onupdate=db.func.now())
