--- app/website/uploads.py
@@ -1,2 +1,0 @@
-
-
