+++ run.py
@@ -1,4 +1,4 @@
-from app.website import create_app
+from app import create_app
 from app.core import core
 import sys
